// src/types/PropertyListing.ts
type PropertyListing = {
  propertyListingId: number;
  nftCollectionId: number;
  nftItemId: number;
  assetId: number;
  id: string;
  blockNumber: number;
  signer: string;
  region: number;
  location: string;
  tokenPrice: number;
  tokenAmount: number;
  propertyId: number;
  addressStreet: string;
  addressTownCity: string;
  area?: string;
  constructionDate?: string;
  estimatedRentalIncome?: number;
  files: string;
  localAuthority?: string;
  map?: string;
  noOfBathrooms?: number;
  noOfBedrooms?: number;
  numberOfTokens?: number;
  offStreetParking?: string;
  outdoorSpace?: string;
  planningPermissionCode?: string;
  postCode?: string;
  pricePerToken: number;
  propertyDescription?: string;
  propertyDevelopmentCode?: string;
  propertyName?: string;
  propertyNumber?: string;
  propertyPrice: number;
  propertyType?: string;
  quality?: string;
  titleDeedNumber?: string;
};

export default PropertyListing;